requirements.txt
amqp==2.2.2
amqplib==1.0.2
anyjson==0.3.3
argh==0.26.2
billiard==3.5.0.3
celery==4.1.0
certifi==2017.11.5
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.8
django-actionkit-client==0.17
django-appconf==1.0.2
django-celery==3.2.2
django-dbtemplates==2.0
django-debug-toolbar==1.9.1
django-dotenv==1.4.1
django-htauth==0.1
django-kombu==0.9.4
django-localflavor==1.6.1
django-localflavor-us==1.1
django-picklefield==1.0.0
djangohelpers==0.19
gdata==2.0.18
gevent==1.2.2
greenlet==0.4.12
gunicorn==19.7.1
idna==2.6
kombu==4.1.0
meld3==1.0.2
MySQL-python==1.2.5
ordereddict==1.1
pathtools==0.1.2
pkg-resources==0.0.0
psycopg2==2.7.3.2
pysqlite==2.8.3
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
supervisor==3.3.3
urllib3==1.22
vine==1.1.4
watchdog==0.8.3
zope.dottedname==4.2
